Move TwoFactorResponse to Entities
